requirements.txt
pandas=2.2.2
Request=2.32.3
streamlit=1.36.0